refactor : marge glsl
three.jsのスタイルに合わせてfragmentとvertexをマージ
