+++ src/contour/ContourMaterial.glsl.ts
@@ -1,21 +1,18 @@
 /**
- * 渦巻き状にUV座標を変形させるシェーダー
+ * 等高線状にテクスチャをマッピングするシェーダー
+ * @see : https://github.com/mrdoob/three.js/blob/master/src/renderers/shaders/ShaderLib/meshphong.glsl.js
  */
-export default () => {
-  // language=GLSL
-  return /* GLSL */ `
+
+// language=GLSL
+export const fragment = /* GLSL */ `
 #define PHONG
 
 #include <mesh_phong_uniform>
 varying vec2 uvPosition;
 #include <mesh_position_varying>
 
-#include <time_animation_uniform_chunk>
-uniform float uvRotation;
-uniform float swirlRotation;
-uniform float radius;
-uniform vec2 center;
-#include <map_uniform_chunk>
+uniform float bottom;
+uniform float top;
 
 #include <common>
 #include <packing>
@@ -43,55 +40,21 @@ uniform vec2 center;
 #include <specularmap_pars_fragment>
 #include <logdepthbuf_pars_fragment>
 #include <clipping_planes_pars_fragment>
-
-/**
- * UV座標を回転する
- */
-vec2 rotateUV(vec2 uv, float rotation, vec2 center)
-{
-  return vec2(
-    cos(rotation) * (uv.x - center.x) + sin(rotation) * (uv.y - center.y) + center.x,
-    cos(rotation) * (uv.y - center.y) - sin(rotation) * (uv.x - center.x) + center.y
-  );
-}
-
-/**
- * UV座標をツイストする
- */
-vec2 swirl(vec2 uv, float radius, float rotation, vec2 center)
-{
-  vec2 tc = uv - center;
-  float dist = length(tc);
-  if (dist < radius) 
-  {
-    float percent = (radius - dist) / radius;
-    float theta = percent * percent * rotation;
-    float s = sin(theta);
-    float c = cos(theta);
-    tc = vec2(dot(tc, vec2(c, -s)), dot(tc, vec2(s, c)));
-  }
-  tc += center;
-  return tc;
-}
-
 void main() {
     #include <clipping_planes_fragment>
   
     #include <mesh_phong_diffuse_color>
     
     #include <logdepthbuf_fragment>
+
     #include <__ShaderMaterial__map_fragment_begin_chunk>
-    #include <map_fragment>
-  
-    mapUV = rotateUV( mapUV, uvRotation , center);
-    mapUV = swirl( mapUV, radius, swirlRotation, center );
-    // offset Texture 
-    mapUV += vec2(time);
-    #include <map_fragment_chunk>
-    
+    #ifdef USE_MAP
+      float mapY = ( meshPosition.y - bottom ) / ( top - bottom );
+      vec4 texelColor = texture2D( map, vec2(0.5, mapY) );
+      diffuseColor *= texelColor;
+    #endif
     #include <color_fragment>
     #include <mesh_phong_switching_alpha_map>
-    
     // #include <alphamap_fragment>
     #include <alphatest_fragment>
     #include <specularmap_fragment>
@@ -114,4 +77,3 @@ void main() {
     #include <premultiplied_alpha_fragment>
     #include <dithering_fragment>
 }`;
-};
