+++ src/crossGrid/CrossGridMaterial.glsl.ts
@@ -1,17 +1,25 @@
-export default () => {
-  // language=GLSL
-  return /* GLSL */ `
+/**
+ * 十字のラインでグリッドを分割するシェーダー
+ *
+ * @see : https://github.com/mrdoob/three.js/blob/master/src/renderers/shaders/ShaderLib/meshphong.glsl.js
+ */
+
+// language=GLSL
+export const fragment = /* GLSL */ `
 #define PHONG
 
 #include <mesh_phong_uniform>
 varying vec2 uvPosition;
 #include <mesh_position_varying>
 
-#include <tiling_fbm_uniform_chunk>
-uniform float progress;
-uniform vec3 edgeColor;
-uniform float edgeWeight;
-#include <tiling_fbm_function_chunk>
+//user settings
+#include <time_animation_uniform_chunk>
+#include <wavy_animation_uniform_chunk>
+#include <repeat_pattern_uniform_chunk>
+#include <mask_map_uniform_chunk>
+#include <reversible_uniform_chunk>
+uniform float gridWeight;
+uniform float radius;
 
 #include <common>
 #include <packing>
@@ -39,8 +47,8 @@ uniform float edgeWeight;
 #include <specularmap_pars_fragment>
 #include <logdepthbuf_pars_fragment>
 #include <clipping_planes_pars_fragment>
-void main()
-{
+
+void main() {
     #include <clipping_planes_fragment>
   
     #include <mesh_phong_diffuse_color>
@@ -49,23 +57,39 @@ void main()
     #include <__ShaderMaterial__map_fragment_begin_chunk>
     #include <map_fragment>
     #include <color_fragment>
+
+    #include <repeat_pattern_fragment_chunk>    
+    vec2 localPos = mod(uv, 1.0) - 0.5;
+    vec2 id = uv - localPos;
+    #include <wavy_animation_fragment_chunk>
+
+    #include <mask_map_fragment_chunk>
+    float w = gridWeight;
+    w = clamp( w, 0.0, 1.0);
     
-    vec2 uv = uvPosition * tiles;
-
-    float fbmVal = fbm(uv);
-    float bri = 1.0 - smoothstep( progress-0.01, progress, fbmVal );
-
-    float edge = 
-          smoothstep( progress-edgeWeight, progress, fbmVal )
-        - smoothstep( progress, progress+edgeWeight, fbmVal );
-    edge = clamp( edge, 0.0, 1.0 );
-
-    vec3 col = diffuseColor.rgb;
-    col += edgeColor * edge;
-
-    diffuseColor.rgb = col;
-    diffuseColor.a *= bri;
+    float margin = clamp ( w * 0.33, 0.00, 0.05 );
     
+    //十字を描画
+    float gridLine;
+    gridLine  = smoothstep ( -w-margin, -w, localPos.x );
+    gridLine -= smoothstep ( w, w+margin, localPos.x );
+    gridLine += smoothstep ( -w-margin, -w, localPos.y );
+    gridLine -= smoothstep ( w, w+margin, localPos.y );
+    gridLine  = clamp( gridLine, 0.0, 1.0 ); 
+
+    //半径でマスク
+    float r = radius - (1.0-mask);
+    gridLine -= smoothstep( r, r+margin, localPos.x);
+    gridLine -= smoothstep( -r, -r-margin, localPos.x);
+    gridLine -= smoothstep( r, r+margin, localPos.y);
+    gridLine -= smoothstep( -r, -r-margin, localPos.y);
+    gridLine = clamp( gridLine, 0.0, 1.0 );
+    
+    gridLine = isReversed
+        ? 1.0 - gridLine
+        : gridLine;
+    diffuseColor.a *= gridLine;
+
     #include <mesh_phong_switching_alpha_map>
 
     // #include <alphamap_fragment>
@@ -90,4 +114,3 @@ void main()
     #include <premultiplied_alpha_fragment>
     #include <dithering_fragment>
 }`;
-};
