+++ src/fbmFire/FBMFireMaterial.glsl.ts
@@ -1,28 +1,27 @@
 /**
- * 6角形グリッドでディゾルブを行うフラグメントシェーダー
- * {@link https://qiita.com/edo_m18/items/37d8773a5295bc6aba3d}
+ * FBMノイズを使った炎の表現
+ *
+ * @see : https://github.com/mrdoob/three.js/blob/master/src/renderers/shaders/ShaderLib/meshphong.glsl.js
  */
-export default () => {
-  // language=GLSL
-  return /* GLSL */ `
+
+// language=GLSL
+export const fragment = /* GLSL */ `
 #define PHONG
 
 #include <mesh_phong_uniform>
 varying vec2 uvPosition;
 #include <mesh_position_varying>
 
-//user settings
-#include <repeat_pattern_uniform_chunk>
-#include <mask_map_uniform_chunk>
-#include <reversible_uniform_chunk>
-uniform float progress;
-uniform float delay;
-uniform float gridWeight;
-uniform bool isAscending;
+#include <tiling_fbm_uniform_chunk>
+#include <tiling_fbm_function_chunk>
+#include <time_animation_uniform_chunk>
 
-uniform vec3 gridEmissive;
-uniform float gridEmissiveWeight;
-#include <hex_grid_function_chunk>
+uniform float strength;
+uniform float bloom;
+
+#include <surface_normal_varying_chunk>
+uniform float rimStrength;
+uniform float rimPow;
 
 #include <common>
 #include <packing>
@@ -50,10 +49,8 @@ uniform float gridEmissiveWeight;
 #include <specularmap_pars_fragment>
 #include <logdepthbuf_pars_fragment>
 #include <clipping_planes_pars_fragment>
-float reverse( float val, bool isReversed){
-  return isReversed ? 1.0 - val : val;
-}
-void main() {
+void main()
+{
     #include <clipping_planes_fragment>
   
     #include <mesh_phong_diffuse_color>
@@ -62,36 +59,37 @@ void main() {
     #include <__ShaderMaterial__map_fragment_begin_chunk>
     #include <map_fragment>
     #include <color_fragment>
-
-    #include <repeat_pattern_fragment_chunk>    
-    vec4 hc = hexCoords( uv );
-    vec2 id = hc.zw;
-
-    #include <mask_map_fragment_chunk>
-  
-    float range = 1.0 - delay;
-    float rateY = isAscending 
-      ? ( division-id.y ) / division
-      : id.y  / division;
-  
-    float currentProgress = progress - (rateY * delay);
-    currentProgress /= range;
-    currentProgress = clamp( currentProgress, 0.0, 1.0);
-  
-    float w = gridWeight + currentProgress / 2.0 + (1.0 - mask);
-    w = clamp( w, 0.0, 1.0);
-    float margin = clamp ( w * 0.33, 0.00, 0.02 );
-  
-    float gridLine = smoothstep(w, w + margin, hc.y);
-    gridLine =  reverse ( gridLine , isReversed);
-    diffuseColor.a *= gridLine ;
     
-    float emmesiveWeight = currentProgress / 2.0 * gridEmissiveWeight;
-    emmesiveWeight =  reverse ( emmesiveWeight, isReversed );
-    float emissiveVal = smoothstep(emmesiveWeight, emmesiveWeight + margin, hc.y);
-    emissiveVal = 1.0 - emissiveVal;
-    diffuseColor.rgb += gridEmissive * emissiveVal;
+    vec2 uv = uvPosition;
+    float uVy = uv.y;
+    uv *= tiles;
 
+    vec2 q = vec2(0.0);
+    q.x = fbm( uv + vec2(1.7,9.2) +.16  * time );
+    q.y = fbm( uv + vec2(8.3,2.8) +.356 * time );
+
+    float fbmVal = fbm(uv + q);
+    fbmVal += 1.0-(uVy * 1.0 );
+    fbmVal *= 1.0-uVy;
+    
+    vec3 viewDir = normalize(vViewPosition);    
+    float rimGlow = 1.0 - max(0.0, dot(surfaceNormal, viewDir));
+    rimGlow = pow(rimGlow, rimPow) * rimStrength;
+    rimGlow = clamp( rimGlow, 0.0, 1.0);
+    fbmVal *= 1.0-rimGlow;
+    
+    vec3 color = diffuseColor.rgb;
+    
+    float st = 1.0 - strength;
+    float bri = smoothstep( max( st - 0.4, 0.0 ), st, fbmVal );
+    
+    float blm = 1.0 - bloom;
+    float bloomVal = smoothstep( blm - 0.4, blm, fbmVal );
+    color += bloomVal;
+
+    diffuseColor.rgb = color;
+    diffuseColor.a *= bri;
+    
     #include <mesh_phong_switching_alpha_map>
 
     // #include <alphamap_fragment>
@@ -116,4 +114,3 @@ void main() {
     #include <premultiplied_alpha_fragment>
     #include <dithering_fragment>
 }`;
-};
