+++ src/hexDissolve/HexDissolveMaterial.glsl.ts
@@ -1,18 +1,30 @@
 /**
- * Cellular Noise Fragment Shader
+ * 6角形グリッドでディゾルブを行うフラグメントシェーダー
+ *
+ * @link : https://qiita.com/edo_m18/items/37d8773a5295bc6aba3d
+ * @see : https://github.com/mrdoob/three.js/blob/master/src/renderers/shaders/ShaderLib/meshphong.glsl.js
  */
-export default () => {
-  // language=GLSL
-  return /* GLSL */ `
+
+// language=GLSL
+export const fragment = /* GLSL */ `
 #define PHONG
 
 #include <mesh_phong_uniform>
 varying vec2 uvPosition;
 #include <mesh_position_varying>
 
-uniform float grid;
-uniform float divisionScaleX;
-#include <time_animation_uniform_chunk>
+//user settings
+#include <repeat_pattern_uniform_chunk>
+#include <mask_map_uniform_chunk>
+#include <reversible_uniform_chunk>
+uniform float progress;
+uniform float delay;
+uniform float gridWeight;
+uniform bool isAscending;
+
+uniform vec3 gridEmissive;
+uniform float gridEmissiveWeight;
+#include <hex_grid_function_chunk>
 
 #include <common>
 #include <packing>
@@ -40,52 +52,9 @@ uniform float divisionScaleX;
 #include <specularmap_pars_fragment>
 #include <logdepthbuf_pars_fragment>
 #include <clipping_planes_pars_fragment>
-
-vec2 rand2D(vec2 p, vec2 scale) {
-    p = mod(p, scale);
-    const float a = 12.9898, b = 78.233, c = 43758.5453;
-    const float a2 = 26.7, b2 = 14.879;
-    
-    highp float dt = dot(p, vec2(a, b)), sn = mod(dt, PI);
-    highp float dt2 = dot(p, vec2(a2, b2)), sn2 = mod(dt2, PI);
-    return fract(sin(vec2(dt, dt2)) * c);
+float reverse( float val, bool isReversed){
+  return isReversed ? 1.0 - val : val;
 }
-
-/*!
- * Cellular Noise
- *
- * The inherits function is :
- * Author : patriciogv
- * see https://thebookofshaders.com/12/
- * LICENSE : https://github.com/patriciogonzalezvivo/thebookofshaders/issues/235
- */
-float cellularNoise(vec2 uv, float grid, float divisionScaleX, float time){
-  
-    vec2 scale = grid * vec2 ( divisionScaleX, 1.0 );
-    uv *= scale;
-    
-    vec2 i_uv = floor(uv);
-    vec2 f_uv = fract(uv);
-    
-    float minDist = 1.;
-    
-    for (int y= -1; y <= 1; y++) {
-        for (int x= -1; x <= 1; x++) {
-            vec2 neighbor = vec2(float(x), float(y));
-            vec2 point = rand2D(i_uv + neighbor, scale);
-            
-            point = 0.5 + 0.5 * sin(time + PI2 * point);
-            
-            vec2 diff = neighbor + point - f_uv;
-            float dist = length(diff);
-            
-            minDist = min(minDist, dist);
-        }
-    }
-    
-    return minDist;
-}
-
 void main() {
     #include <clipping_planes_fragment>
   
@@ -96,10 +65,35 @@ void main() {
     #include <map_fragment>
     #include <color_fragment>
 
-    float dist = cellularNoise( mapUV, grid, divisionScaleX, time );
-    diffuseColor.rgb *= dist;
-    diffuseColor.a *= dist;
+    #include <repeat_pattern_fragment_chunk>    
+    vec4 hc = hexCoords( uv );
+    vec2 id = hc.zw;
+
+    #include <mask_map_fragment_chunk>
+  
+    float range = 1.0 - delay;
+    float rateY = isAscending 
+      ? ( division-id.y ) / division
+      : id.y  / division;
+  
+    float currentProgress = progress - (rateY * delay);
+    currentProgress /= range;
+    currentProgress = clamp( currentProgress, 0.0, 1.0);
+  
+    float w = gridWeight + currentProgress / 2.0 + (1.0 - mask);
+    w = clamp( w, 0.0, 1.0);
+    float margin = clamp ( w * 0.33, 0.00, 0.02 );
+  
+    float gridLine = smoothstep(w, w + margin, hc.y);
+    gridLine =  reverse ( gridLine , isReversed);
+    diffuseColor.a *= gridLine ;
     
+    float emmesiveWeight = currentProgress / 2.0 * gridEmissiveWeight;
+    emmesiveWeight =  reverse ( emmesiveWeight, isReversed );
+    float emissiveVal = smoothstep(emmesiveWeight, emmesiveWeight + margin, hc.y);
+    emissiveVal = 1.0 - emissiveVal;
+    diffuseColor.rgb += gridEmissive * emissiveVal;
+
     #include <mesh_phong_switching_alpha_map>
 
     // #include <alphamap_fragment>
@@ -124,4 +118,3 @@ void main() {
     #include <premultiplied_alpha_fragment>
     #include <dithering_fragment>
 }`;
-};
