+++ src/rimEffect/OuterGlowMaterial.ts
@@ -1,6 +1,10 @@
 import { IExpandable } from "../chunk/";
 import { RimEffectMaterial } from "./RimEffectMaterial";
 
+/**
+ * モデルの外周を発光させるマテリアル。
+ * 縁の発光と膨張を掛け合わせて、元のモデルの周辺を発光させる。
+ */
 export class OuterGlowMaterial
   extends RimEffectMaterial
   implements IExpandable
