+++ src/squareGrid/SquareGridMaterial.ts
@@ -1,6 +1,6 @@
 import VertexShader from "../ShaderPhongMaterial.vert.glsl";
 import { WavyGridMaterial } from "../WavyGridMaterial";
-import FragmentShader from "./SquareGridMaterial.frag.glsl";
+import { fragment } from "./SquareGridMaterial.glsl";
 import { ShaderMaterialParameters, UniformsUtils } from "three";
 
 /**
@@ -19,7 +19,7 @@ export class SquareGridMaterial extends WavyGridMaterial {
   }
 
   constructor(parameters?: ShaderMaterialParameters) {
-    super(VertexShader(), FragmentShader(), parameters);
+    super(VertexShader(), fragment, parameters);
   }
 
   protected initUniforms(): void {
